FEM: fem command package, rename it to commands
